refactoring according to comments in pull/202
